erlang-cloudi-service: Make the return of impl handlers conforming to the cloudI interface
Make handler functions (cloudi_service_handle_request) in impl module to
return conforming to the cloudI interface (e.g., {reply, ResponseInfo,
Response, StateNew}). The primary reason for this is to make it
consistent with the C++ generator. Also, this makes this funcions to
conform with cloudI interface, so, if we validation layer is removed,
these handler in the impl could be invoked directly.  # Please enter the
commit message for your changes. Lines starting

Change-Id: Ib8530ed8f29918c1fc596bb759760016fc62a795